+++ tool/bots.sh
@@ -155,30 +155,35 @@ elif [[ "$BOT" == "test_ddc" || "$BOT" == "test_dart2js" ]]; then
 # elif [ "$BOT" = "integration_ddc" ]; then
 
 # TODO(https://github.com/flutter/devtools/issues/1987): rewrite legacy integration tests.
-elif [ "$BOT" = "dart2js" ]; then
+elif [ "$BOT" = "integration_dart2js" ]; then
+    if [ "$DEVTOOLS_PACKAGE" = "devtools_app" ]
+        flutter pub get
 
-    flutter pub get
+        # TODO(https://github.com/flutter/flutter/issues/118470): remove this warning.
+        echo "Preparing to run integration tests.\nWarning: if you see the exception \
+    'Web Driver Command WebDriverCommandType.screenshot failed while waiting for driver side', \
+    this is a known issue and likely means that the golden image check failed (see \
+    https://github.com/flutter/flutter/issues/118470). Run the test locally to see if new \
+    images under a 'failures/' directory are created as a result of the test run:\n\
+    $ dart run integration_test/run_tests.dart --headless"
 
-    # TODO(https://github.com/flutter/flutter/issues/118470): remove this warning.
-    echo "Preparing to run integration tests.\nWarning: if you see the exception \
-'Web Driver Command WebDriverCommandType.screenshot failed while waiting for driver side', \
-this is a known issue and likely means that the golden image check failed (see \
-https://github.com/flutter/flutter/issues/118470). Run the test locally to see if new \
-images under a 'failures/' directory are created as a result of the test run:\n\
-$ dart run integration_test/run_tests.dart --headless"
+        if [ "$DEVICE" = "flutter" ]; then
 
-    if [ "$DEVICE" = "flutter" ]; then
+            dart run integration_test/run_tests.dart --headless --shard="$SHARD"
 
-        dart run integration_test/run_tests.dart --headless --shard="$SHARD"
+        elif [ "$DEVICE" = "flutter-web" ]; then
 
-    elif [ "$DEVICE" = "flutter-web" ]; then
+            dart run integration_test/run_tests.dart --test-app-device=chrome --headless --shard="$SHARD"
 
-        dart run integration_test/run_tests.dart --test-app-device=chrome --headless --shard="$SHARD"
+        elif [ "$DEVICE" = "dart-cli" ]; then
 
-    elif [ "$DEVICE" = "dart-cli" ]; then
+            dart run integration_test/run_tests.dart --test-app-device=cli --headless --shard="$SHARD"
 
-        dart run integration_test/run_tests.dart --test-app-device=cli --headless --shard="$SHARD"
-
+        fi        
+    elif [ "$DEVTOOLS_PACKAGE" = "devtools_extensions" ]
+        popd
+        pushd packages/devtools_extensions
+        dart run integration_test/run_tests.dart --headless
     fi
 fi